flathub-data-checker: Initial commit
Add global variant of external-data-checker which clones entire Flathub
organization and runs the tool on all manifests. This is workaround
for GitHub missing API for defining secrets for repositories and
a way to define an action for all repositories in organization.

At the moment it performs a dry run printing status on stdout, after
some testing it'll be changed to use --update switch.